pyproject.toml
[project]
name = "drf-learning-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (==5.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ipython (>=9.10.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
    | \.venv
    | venv
)/'''

[tool.isort]
line_length = 119
